+++ src/gui/widgets/widget.py
